Add conditional rendering to edit / delete review buttons

+++ frontend/components/reviews/edit_review_buttons.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { withRouter } from 'react-router-dom';
+import { logoutCurrentUser } from '../../actions/session_actions';
 
 class EditReviewButton extends React.Component {
   constructor(props) {
@@ -8,14 +9,22 @@ class EditReviewButton extends React.Component {
 
   render() {
     const { review, deleteReview } = this.props
-    
-    return (
-      <div className="review-index-item">
-        <button onClick={() => deleteReview(review)} className="delete-review">Delete Review</button>
-        <button className="update-review">Update Review</button>
-      </div>
-    )
+    const currUser = this.props.currUser === undefined ? {id: null} : this.props.currUser
+
+    if (review.user_id === currUser.id) {
+      return (
+        <div className="review-index-item">
+          <button onClick={() => deleteReview(review)} className="delete-review">Delete Review</button>
+          <button className="update-review">Update Review</button>
+        </div>
+      )
+    } else {
+      return (
+        <div className="hidden">
+        </div>
+      )
+    }
   }
 }
-
+ 
 export default EditReviewButton;
